+++ src/BackupPolicy.h
@@ -25,6 +25,7 @@
 
 class Volume;
 class Device;
+class PolicyParameter;
 
 /** @brief Base class for backup policies
  */
@@ -45,15 +46,16 @@ public:
    * @param volume Volume to get parameter from
    * @param name Name of parameter
    */
-  const std::string &get(const Volume *volume, const std::string &name) const;
+  const PolicyParameter &get(const Volume *volume,
+                             const std::string &name) const;
 
   /** @brief Get a parameter value
    * @param volume Volume to get parameter from
    * @param name Name of parameter
    * @param def Default value
    */
-  const std::string &get(const Volume *volume, const std::string &name,
-                         const std::string &def) const;
+  const PolicyParameter get(const Volume *volume, const std::string &name,
+                            const std::string &def) const;
 
   /** @brief Find a backup policy by name
    * @param name Name of policy