+++ src/PrunePolicy.h
@@ -26,6 +26,7 @@
 
 class Backup;
 class Volume;
+class PolicyParameter;
 
 /** @brief Base class for pruning policies
  */
@@ -46,15 +47,16 @@ public:
    * @param volume Volume to validate
    * @param name Name of parameter
    */
-  const std::string &get(const Volume *volume, const std::string &name) const;
+  const PolicyParameter &get(const Volume *volume,
+                             const std::string &name) const;
 
   /** @brief Get a parameter value
    * @param volume Volume to validate
    * @param name Name of parameter
    * @param def Default value
    */
-  const std::string &get(const Volume *volume, const std::string &name,
-                         const std::string &def) const;
+  const PolicyParameter get(const Volume *volume, const std::string &name,
+                            const std::string &def) const;
 
   /** @brief Identify prunable backups
    * @param onDevice Surviving backups of same volume on same device